rom: Move checksum to cartridge, pass cartridge instead of vector

+++ gamekid/utils/offset.h
@@ -2,8 +2,8 @@
 
 namespace gamekid::utils {
     struct offset {
-        int start;
-        int length;
+        size_t start;
+        size_t length;
 
         offset(int start, int length) : start(start), length(length) {}
     };
